+ Add skeleton projects for PS4 + Make sure the game layer compiles under orbis-clang

+++ src/platform.h
@@ -117,7 +117,7 @@ typedef int64_t i64;
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
-typedef uint64_t u64;
+typedef unsigned long long u64;
 
 typedef float r32;
 typedef double r64;
